# Avoid re-using RE2 regex between threads, since RE2 regex objects have locks internally