This greatly improves performance of regex functions in the presence of multiple threads:


```sql
select count(*) from lineitem where l_shipinstruct SIMILAR TO '.*DE.*I.*ER.*';
```

| Branch | 1 Thread | 4 Threads |
|--------|----------|-----------|
| Old    | 0.64s    | 1.26s     |
| New    | 0.64s    | 0.2s     |